feat: introduce LoginViewModel for login screen
